Pesquisa deu certo

+++ php/listar_livro.php
@@ -34,7 +34,7 @@
                 'titulo',
                 'autor',
                 'capa',
-                'capa2',
+                'artigo',
                 'pdf',
                 'banca',
                 'texto'
@@ -66,9 +66,9 @@
             <th class="col">Autor</th>
             <th class="col">Banca</th>
             <th class="col">Resumo</th>
+            <th class="col">Artigo</th>
             <th class="col">PDF</th>
             <th class="col">Capa</th>
-            <th class="col">Segunda Capa</th>
             </tr>
         </thead>
         <tbody>
@@ -77,12 +77,12 @@
                 <td><?php echo $entidade['autor'] ?></td>
                 <td><?php echo $entidade['banca'] ?></td>
                 <td><?php echo $entidade['texto'] ?></td>
+                <td><?php echo $entidade['artigo'] ?></td>
                 <td><a href="<?php echo $pasta_pdf.$entidade['pdf'] ?>">PDF</a></td>
                 <td><img src="<?php echo $pasta.$entidade['capa']?>" width="200px" height="400px"> </td>
-                <td><img src="<?php echo $pasta.$entidade['capa2']?>" width="200px" height="400px"></td>
                 <!--<td><a href="../core/livro_repositorio.php?acao=delete&idLivro=<?php //echo $entidade['idLivro']?>">Deletar</a></td>-->
-                <td><a href="excluir_livro.php?idLivro=<?php echo $entidade['idLivro']?>">Deletar</a></td>
-                <td><a href="../cadastro_livro.html?idLivro=<?php echo $entidade['idLivro']?>">Atualizar</a></td>
+                <td><a href="excluir_livro.php?idLivro=<?php echo $entidade['idLivro']?>" onclick="return confirm('Deseja excliur permanentemente o livro?');">Deletar</a></td>
+                <td><a href="../cadastro_livro.php?idLivro=<?php echo $entidade['idLivro']?>">Atualizar</a></td>
             <?php endforeach ?>
         </tbody>
     </table>
